% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQP_auto.R
\name{SQP_auto}
\alias{SQP_auto}
\title{Title}
\usage{
SQP_auto(
  init,
  f_exp,
  h_exp,
  xdim = 2,
  meq = 0,
  method = "standard",
  max_iter = 10000
)
}
\arguments{
\item{max_iter}{}
}
\value{

}
\description{
Title
}
